Allow autoscrolling to continue when the cursor leaves the page.
A CL[1] over a year ago added an if check that caused autoscrolling to stop
if the cursor left the page and the middle button wasn't held in. Since
other Windows apps allow auto scrolling to continue in this case, it
makes sense to match that behavior and also allow it.

[1] https://crrev.com/5971e47dc08dbb773f58fd6fb13b1df7917fd295

Bug: 806562
Change-Id: Id7259b5242f21c2042f300d7f0a606179be89b5c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1618141
Commit-Queue: Mario Bianucci <mabian@microsoft.com>
Reviewed-by: David Bokan <bokan@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#662184}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4849af4790babb032dcd87718f41782272004559