[UserTimingL3] Throw error when global object is not Window object
The spec[1] says "If the global object is not a Window object, throw
a TypeError". We should check whether the global object is a window
object when converting a performance-timing name to timestamp. If
it is not, throw a Type Error.

[1] https://w3c.github.io/user-timing/#convert-a-name-to-a-timestamp

Bug: 953869
Change-Id: I188d810f2505a37f2cdfc823ce076cec96dd7839
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1613653
Commit-Queue: Liquan (Max) Gu <maxlg@chromium.org>
Reviewed-by: Nicolás Peña Moreno <npm@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#662216}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9faed217f0e81084630520368cc4a028686a6437