DL: Don't process update phases if we're not in a lifecycle run.
This patch ensures that we don't process any update() lifecycle phases
when layout is forced outside of the lifecycle phase. This can happen
if the budget yields, and then script runs which forces something
to be forced. We don't want to process update phases in that case,
since we want to wait until proper lifecycle runs to do that.

R=chrishtr@chromium.org, rakina@chromium.org

Bug: 964827
Change-Id: I05848c9dd3dcd4fcedc209a2f66e324d2ba6f0b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1626172
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Commit-Queue: vmpstr <vmpstr@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#662612}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6275585dd2210038fac23b0f3b86f94a6ff9d8c3