Fix crash when HTMLPreloadScanner encounters atypical picture children
When a picture element was opened but not closed, internal tags would
get preloaded as image_set typed resources, which shouldn't happen when
the loaded resources are not images.

This CL fixes that by terminating picture-like processing as soon as an
atypical tag is encountered inside `<picture>`.
That may reduce preloading for invalid `<picture>` tags, but that is
most-probably fine.

Bug: 961151
Change-Id: I392b87e51100175b38461d47f7677c840448c78d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1626590
Commit-Queue: Yoav Weiss <yoavweiss@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#662674}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c1f636e0e3e9f51cf90195570b561bb97a478e38