DL: Allow child to need style recalc after rebuilding layout tree
It's possible to call RebuildLayoutTree on a locked element with a set
ChildNeedsStyleRecalc flag, because we are blocking child style recalc
but not self style recalc on that locked element.

This CL modifies the DCHECK in Element::RebuildLayoutTree to expect
that case.

Bug: 966310
Change-Id: I53338285df4d51d04ff5e92e41be5e928b0ac154
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1627272
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Reviewed-by: vmpstr <vmpstr@chromium.org>
Commit-Queue: Rakina Zata Amni <rakina@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#662718}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c50dd2f5725db68b89e470b4968c7d4bd8745b04