DL: Prevent composited layers from painting when display-locked.
This patch ensures that we don't paint composited layers when display
locked. This is similar to how we block non-composited layers as well.

The reason for this is that they may not be laid out, and the display
locking implemention (for now) ensures that nothing is painted.

In the future this will change to paint just the self-element, not
the subtree.

R=chrishtr@chromium.org, rakina@chromium.org

Bug: 966310
Change-Id: I8d5e557d833e0757e68bb5467224c0c74619e389
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1627550
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Commit-Queue: vmpstr <vmpstr@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#662750}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3e0919da0d715622e8ef819bccc9a097c613b36d