Make WebXR Gamepad y axis values consistent.
According to the Gamepad spec, y axis values should have -1.0 be up and
1.0 be down. WMR, OpenVR, and Oculus device APIs have 1.0 correspond to
up, so we have to flip the signs for the y axis values that come from
those platforms. The Oculus gamepad builder code already flips the sign
for the y axis, which is why it's not included in this CL. Note that
this will also affect the WebVR gamepads in the same way.

Changes to input-selection.html are to help testing and debugging WebXR
gamepads. Now the page shows tables for button and axis values which are
updated in real-time.

Bug: 965522, 965527
Change-Id: I08a1ecf4d5f90f41edcf5574fc9fd96e043482a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1626393
Reviewed-by: Bill Orr <billorr@chromium.org>
Reviewed-by: Klaus Weidner <klausw@chromium.org>
Reviewed-by: Alexander Cooper <alcooper@chromium.org>
Commit-Queue: Jacob DeWitt <jacde@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#662772}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c1ae7c6ced32ccbed12548962eeda33d34d92e50