[WebAudio DevTools] Flush context map when the frame is loaded
When a page/frame is refreshed, some BaseAudioContexts from the previous
page stay in WebAudioModel object. This is because of a flaky behavior
of BaseAudioContext::ContextDestroyed().

To make it clear and consistent, this change will flush WebAudioModel
whenever a page/frame is loaded.

Bug: 965310
Test: Locally confirmed the view panel is updated when the frame is loaded.
Change-Id: I6e35dc0694e743f80ee063e76b0b9affc74ea75f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1626531
Commit-Queue: Hongchan Choi <hongchan@chromium.org>
Reviewed-by: Erik Luo <luoe@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#662805}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c1e88cf742bad768e94857ac3343f2778d731f2b