[LayoutNG] Ignore fragmentation properties when not fragmenting.
When we're not fragmenting, we shouldn't touch any of the fragmentation
related values in the container builder.

Added DCHECKs that would fail without this fix (and that will catch any
similar problems in the future at an earlier stage).

Bug: 966369
Change-Id: I8f6300435e211086fea8ea381912ad595e9a91ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1628707
Commit-Queue: Emil A Eklund <eae@chromium.org>
Reviewed-by: Emil A Eklund <eae@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#663051}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 244f618056cddb29ae2a9ef3540746ddedd359d8