Deflake fixed-background-in-iframe web test
crrev.com/c/1629054 modified how scrollbars scroll, under a feature
flag. Originally this wasn't under a flag and needed an update to this
test to get it to pass reliably, due to how the timing of the scroll
animation setup had changed (namely with the new implementation the main
thread has to yield before the injected gestures will trigger the scroll
animation).

The modification to the test passed locally and on trybots even with the
extra rAF. However per FindIt, this caused the test to get flaky since
the animation can complete before the second rAF fires if the machine
is under load.

I was able to repro this locally if I ran the test while compiling. To
fix this, the test now passes if the main thread scroll animation is
running on either the first or second rAF.

Change-Id: Ia7ebcfe092addb87a42e6d6ce7e638b78ef639cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1630640
Reviewed-by: David Bokan <bokan@chromium.org>
Commit-Queue: Daniel Libby <dlibby@microsoft.com>
Cr-Original-Commit-Position: refs/heads/master@{#663600}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cf80b560ed7f987abd905527bcdf773f4560a2d1