[LayoutNG] Fix NeedsCollectInlines tp stop at atomic inlines
This patch stops propagation of setting NeedsCollectInlines
at atomic inlines.

Usually atomic inlines that have children are LayoutBlockFlow,
but LayoutSVGRoot is one example that are not LayoutBlockFlow.

This patch also clarifies the distinction when a parent of
object needs to collect from when the object itself needs to
collect.

Bug: 966699
Change-Id: I260824f3e8e211f842aa7990c6d504a3c72549b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1631016
Commit-Queue: Emil A Eklund <eae@chromium.org>
Reviewed-by: Emil A Eklund <eae@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#663760}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2146d7c40f5122395d26ef3e25be4e63d228a792