Media Controls: Prevent overflow from closing on a click and drag
This CL adds a check in the popup menu click handling to only assume an
item was selected when the click target is not the popup menu itself.
This fixes an issue where the popup would close when a user clicked and
dragged between two items (since this sends a "click" event with the
parent as the target).

Bug: 934225
Change-Id: I67b7079b0a8a0d1d6405953808f805cfed0c83ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1632421
Reviewed-by: Jazz Xu <jazzhsu@chromium.org>
Commit-Queue: Tommy Steimel <steimel@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#663846}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 900c55429fbf0236f7b6f3d8e94d2bd68066ca1d