[WebXR] Ensure Gamepad.connected attribute is set false on disconnect
When an input_source is marked as removed, its gamepad needs to also be
marked as disconnected.  To facilitate writing tests for this, the blink
web_test mocks were also updated to allow for connecting/disconnecting a
gamepad object.

Also includes very minor cleanup to the test that was used as a pattern
when creating the newly added web_test.

Bug: 963013
Change-Id: Ia985b548dbc6b1e7052a6189a484b6a93298e2b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1626885
Reviewed-by: Bill Orr <billorr@chromium.org>
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#663917}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 842fb66792b3f6eabcdf090b8a5e2c9fd75cc4de