Avoid triggering spurious downloads for no-store responsive images
When responsive images are not cached in the memory cache (due to
`no-store` or devtools settings) and the viewport is being resized,
that triggers spurious downloads.

This CL fixes that, by aligning to the spec (somewhat) and making sure
that viewport changes that don't modify the image URL do not trigger a
resource load.

Bug: 960633
Change-Id: I8105bcf5dbef99767ac704872800f4d417caf8b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1622168
Reviewed-by: Fredrik Söderquist <fs@opera.com>
Commit-Queue: Yoav Weiss <yoavweiss@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#664269}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1e5cba0e1d5af735e1b980fd62b5e24b61735267