a11y: Add popup children for nodes walked via layout tree
Call AddPopupChildren() when walking the layout tree in order to ensure
that popup contents for controls such as <input type="date"> are added
to the accessibility tree.

For testing, I wanted a tree dump that would show the contents of the
popup. I tried a couple of different techniques to simulate a click on
the popup button from script but none of them worked. Instead I added
the ability for the tree dumper to synthesize clicks on elements after
the document is loaded.

Bug: 964154
Change-Id: I75b6169bfd336bd98c70e93960f4f56a6413424f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1616387
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Ian Prest <iapres@microsoft.com>
Commit-Queue: Kevin Babbitt <kbabbitt@microsoft.com>
Cr-Original-Commit-Position: refs/heads/master@{#664435}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 700f882cc2a227773c06d7b260dfdd5a34ba4164