Fix nullptr dereference in WebAXObject
A WebAXObject holds  An AXObject as a WebPrivatePtr which holds a AXObjectCacheImpl Member.

In order to persist the AXObjectCacheImpl, we need to directly hold a Persistent<AXObjectCacheImpl> in ScopedActionAnnotator so that in our destructor, we can be guaranteed the AXObjectCacheImpl is still around.

Bug: 966935
Change-Id: I5c0c1322af58d5ee388e3ef448c0ccf767e50b34
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1629184
Commit-Queue: David Tseng <dtseng@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#664588}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4f15d4fabbdf9c150deafb39d5424804cb25501e