DL: Force style update for elements in a locked subtree in IsFocusableStyle
Callers to IsKeyboardFocusable & IsFocusableStyle are guaranteed to call
with style-clean for the document, except for elements within a locked
subtree, because they are skipped in normal style recalc traversals.

This CL makes sure we force the style recalc for elements in a locked
subtree in IsFocusableStyle.

Bug: 966483
Change-Id: I576604851accac118a3569eddeb129660b61e38d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1636989
Reviewed-by: Kent Tamura <tkent@chromium.org>
Commit-Queue: Rakina Zata Amni <rakina@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#664641}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bdad5fe9550437a9ee3771cc6c1d1470716671f8