[XR] Ensure that selectend and selectstart events are always paired
Per the spec, a selectend event should always be sent if a selectstart
event has been fired, evenif the gesture is cancelled.  Currently if an
input source was removed but it had sent a selectstart event a selectend
event would never be sent.  This fixes that for both the cases where an
input source is lost, and if the session ends while a controller is
pressed.

Per the HTML interaction spec, since these removals cannot be
guaranteed to have fired as a result of user interaction, the frame
is not notified of UserActivation in this case (and as such the
OnSelectEnd method is not re-used).

Bug: 968291
Change-Id: Ibea99e59ee7a528b7a34db610bfca76c27d495b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1636851
Reviewed-by: Bill Orr <billorr@chromium.org>
Reviewed-by: Brandon Jones <bajones@chromium.org>
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#664805}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cb872f62d1939dd29871068c0be258e45e99c6bf