IndexedDB: Record IDBRequest metrics after dispatch has occurred
Tracing metrics are captured when IDBRequest handles a response
that effectively completes a callback or action in IndexedDB.
Previously, these metrics were marked as done when the response
was enqueued which meant the metric didn't account for the time
the event needed to be fully dispatched and run.

This change updates the metrics so they are recorded once they
are dispatched.  The result will be that after this change the
metrics will all show a regression but in reality they are only
being made more accurate now.

We expect a future change that directly dispatches events to
improve these metrics.

Bug: 968536
Change-Id: I1801aa37e6946073f1ecabf795c0d9763bc413d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1637304
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Commit-Queue: Chase Phillips <cmp@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#665001}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c9dc17886aa6892207c2772fcc58aa02f088a408