OnionSoup: Remove recent new usages of ScopedMockClock
This is a follow up of crrev.com/c/1518518 in which 3 new usages of
ScopedMockClock are replaced by the use of TestMockTimeTaskRunner. Handling
these in a different CL because the parent one is already big enough to review.

Bug: 919383
Change-Id: I8cdeb076ef0b7104a72f340eaccb5d2047886d46
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1593252
Commit-Queue: Sergio Villar <svillar@igalia.com>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#665097}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 01976bc22a3576df9d43c2698a0137ffa77a428d