Fix firing of finished event for animations with an end delay.
If an animation has a positive-valued end delay, we need an additional
tick at the end time to ensure that the finished event is delivered.
Otherwise we stop ticking the animation at the end of the active phase
and the limit is not reached until triggered by another event such as
a mouse move|click.

Bug: 840637
Change-Id: Id811c787833a34f8739b2c10dc2af5b19fa4d17b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1634387
Reviewed-by: Stephen McGruer <smcgruer@chromium.org>
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#665135}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 83ed3e32ee4432d9d03ee8eabe89896dfc8d5b5b