Remove MutableStyle from dark-mode unittests.
I don't believe(?) there is actually a way to set the opacity on a
viewport. (Either with the @viewport rule, or the style fixup[1]).

This removes one of the unittests, and removes the logic which
was reading the opacity from the viewport.

[1] https://cs.chromium.org/chromium/src/third_party/blink/renderer/core/dom/document.cc?q=viewport_style&sq=package:chromium&dr=C&l=2122

Bug: 813068
Change-Id: I29d1b45c848c3ddca48a04523ec358094e97755b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1637044
Reviewed-by: Aran Gilman <gilmanmh@google.com>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#665148}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 039da5b7aeface0af72e4f4c71ea02c0301a157e